Port to SDL 3

+++ src/Headers/mousesmoothing.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <SDL.h>
+#include <SDL3/SDL.h>
 
 void MouseSmoothing_ResetState(void);
 
@@ -8,4 +8,4 @@ void MouseSmoothing_StartFrame(void);
 
 void MouseSmoothing_OnMouseMotion(const SDL_MouseMotionEvent* motion);
 
-void MouseSmoothing_GetDelta(int* dxOut, int* dyOut);
+void MouseSmoothing_GetDelta(float* dxOut, float* dyOut);
